2015/src/test/java/me/zodac/advent/Day04Test.java
/*
 * BSD Zero Clause License
 *
 * Copyright (c) 2021-2022 zodac.me
 *
 * Permission to use, copy, modify, and/or distribute this software for any
 * purpose with or without fee is hereby granted.
 *
 * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
 * WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
 * MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY
 * SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
 * WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
 * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR
 * IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 */

package me.zodac.advent;

import static org.assertj.core.api.Assertions.assertThat;

import java.security.NoSuchAlgorithmException;
import org.junit.jupiter.api.Test;

/**
 * Tests to verify answers for {@link Day04}.
 */
class Day04Test {

    private static final String PUZZLE_INPUT = "yzbqklnj";

    @Test
    void part1() throws NoSuchAlgorithmException {
        final long index = Day04.iterateHashesToFindPrefix(PUZZLE_INPUT, "00000");
        assertThat(index)
            .isEqualTo(282_749L);
    }

    @Test
    void part2() throws NoSuchAlgorithmException {
        final long index = Day04.iterateHashesToFindPrefix(PUZZLE_INPUT, "000000");
        assertThat(index)
            .isEqualTo(9_962_624L);
    }
}
